app/requirements.txt
matplotlib==3.5.1
numpy==1.21.6
pandas==1.3.5
Pillow==9.2.0
seaborn==0.11.2
streamlit==1.11.0
tensorflow==2.9.1
tensorflow_macos==2.8.0
